refactor code and delint stuff
