+++ atm_to_cam/constants.py
@@ -11,13 +11,13 @@ kappa = Rd / cp
 vert_interp_thresh = 0.1  # ps corr diff (Pa) req. to interp vert profiles
 extrap_threshold = 5000.  # maximum ps corr diff (Pa) to allow extrapolation
 
-NC_FLOAT_FILL=9.96921e+36
+NC_FLOAT_FILL = 9.96921e+36
 
 dtime_map = [4, 2, 2, 2]
 
-QMAXTHRESH=0.05
-QMINTHRESH=1.0e-12
-CLDMINTHRESH=0.0
+QMAXTHRESH = 0.05
+QMINTHRESH = 1.0e-12
+CLDMINTHRESH = 0.0
 OMEGA_LAT_THRESH = 88.0
 
 rho_d_algo = 1
@@ -25,9 +25,9 @@ rho_d_algo = 1
 damp_upper_winds_mpas = False
 mpas_uv_damping_coeffs = [0.90, 0.95, 0.98]
 
-ps_wet_to_dry=False
-output_diag=True
+ps_wet_to_dry = False
+output_diag = True
 
-w_smooth_iter=0
+w_smooth_iter = 0
 
-t_freeze_K=273.15
+t_freeze_K = 273.15
