+++ atm_to_cam/vertremap.py
@@ -1,63 +1,87 @@
-import numpy as np
-#from scipy.interpolate import interp1d
-from numba import jit
-#from tqdm import tqdm
 import time
+import logging
+import warnings
 
-# def pressure_to_hybrid(p_levels, data_on_p_levels, ps, a_coeff, b_coeff, p0=100000):
-#     """
-#     Interpolate data on constant pressure levels to hybrid sigma levels.
-#
-#     Parameters:
-#     p_levels: 1D array of pressure levels (Pa) on which data is currently defined
-#     data_on_p_levels: 3D array of data defined on pressure levels (e.g., T, U, V)
-#     ps: 2D array of surface pressure (Pa)
-#     a_coeff: 1D array of a(k) hybrid coefficients
-#     b_coeff: 1D array of b(k) hybrid coefficients
-#     p0: Reference pressure, typically 100000 Pa (1000 hPa)
-#
-#     Returns:
-#     data_on_hybrid_levels: 4D array of data interpolated to hybrid sigma levels
-#     """
-#     start_time = time.time()
-#
-#     # Compute hybrid pressure levels
-#     hybrid_p = np.array([a * p0 + b * ps for a, b in zip(a_coeff, b_coeff)])
-#
-#     # Prepare an output array for the interpolated data
-#     nlevs = len(a_coeff)
-#     nlat, nlon = ps.shape
-#     data_on_hybrid_levels = np.zeros((nlevs, nlat, nlon))
-#
-#     # Interpolate for each lat/lon point with progress bar
-#     for i in tqdm(range(nlat), desc="Interpolating", unit="lat"):
-#         for j in range(nlon):
-#             f_interp = interp1d(p_levels, data_on_p_levels[:, i, j], fill_value="extrapolate")
-#             data_on_hybrid_levels[:, i, j] = f_interp(hybrid_p[:, i, j])
-#
-#     # Print elapsed time
-#     elapsed_time = time.time() - start_time
-#     print(f"Interpolation completed in {elapsed_time:.2f} seconds.")
-#
-#     return data_on_hybrid_levels
+import numpy as np
+from numba import jit
 
-#### NUMBA
+logger = logging.getLogger(__name__)
 
-# @jit(nopython=True)
-# def pressure_to_hybrid_numba(p_levels, data_on_p_levels, ps, hybrid_p):
-#     """
-#     JIT-compiled function to interpolate data on constant pressure levels to hybrid sigma levels.
-#     This function is optimized with numba.
-#     """
-#     nlevs, nlat, nlon = hybrid_p.shape
-#     data_on_hybrid_levels = np.zeros((nlevs, nlat, nlon))
-#
-#     # Interpolate for each lat/lon point
-#     for i in range(nlat):
-#         for j in range(nlon):
-#             data_on_hybrid_levels[:, i, j] = np.interp(hybrid_p[:, i, j], p_levels, data_on_p_levels[:, i, j])
-#
-#     return data_on_hybrid_levels
+__pres_lev_mandatory__ = np.array([
+    1000, 925, 850, 700, 500, 400, 300, 250, 200, 150, 100, 70, 50, 30, 20, 10,
+    7, 5, 3, 2, 1
+]).astype(np.float64) * 100.0  # Convert mb to Pa
+
+
+def interpolate_to_levels(ppin, xxin, ppout, linlog=1, xmsg=np.nan):
+    """
+    Interpolate or extrapolate data from one set of pressure levels to another.
+
+    Parameters:
+    -----------
+    ppin : numpy.ndarray
+        Input pressure levels (1D array).
+    xxin : numpy.ndarray
+        Input variable values at ppin levels (1D array).
+    ppout : numpy.ndarray
+        Output pressure levels (1D array).
+    linlog : int, optional
+        Flag indicating interpolation method:
+        - 1: linear interpolation (default)
+        - 0: logarithmic interpolation
+        - Negative: extrapolation allowed using the nearest valid slope.
+    xmsg : float, optional
+        Missing data marker. Default is np.nan.
+
+    Returns:
+    --------
+    xxout : numpy.ndarray
+        Interpolated variable values at ppout levels.
+    """
+
+    # Initialize output array with missing values
+    xxout = np.full_like(ppout, xmsg)
+
+    # Error check: make sure we have enough points
+    if len(ppin) < 2 or len(ppout) < 1:
+        return xxout  # Return all missing values
+
+    # Determine if input and output pressures need to be reordered
+    ppin_reordered = ppin[::-1] if ppin[0] < ppin[1] else ppin
+    xxin_reordered = xxin[::-1] if ppin[0] < ppin[1] else xxin
+    ppout_reordered = ppout[::-1] if ppout[0] < ppout[1] else ppout
+
+    # Filter out missing data in xxin
+    valid_mask = (xxin_reordered != xmsg) & (ppin_reordered != xmsg)
+    pin_valid = ppin_reordered[valid_mask]
+    xin_valid = xxin_reordered[valid_mask]
+
+    if len(pin_valid) < 2:
+        return xxout  # Not enough valid data points
+
+    # Perform interpolation
+    if linlog == 1:
+        # Linear interpolation
+        xxout = np.interp(ppout_reordered, pin_valid, xin_valid)
+    else:
+        # Logarithmic interpolation
+        xxout = np.interp(np.log(ppout_reordered), np.log(pin_valid), xin_valid)
+
+    # Extrapolation if linlog < 0
+    if linlog < 0:
+        if ppout_reordered[0] > pin_valid[0]:  # Extrapolate above range
+            slope = (xin_valid[1] - xin_valid[0]) / (pin_valid[1] - pin_valid[0])
+            xxout[ppout_reordered > pin_valid[0]] = xin_valid[0] + slope * (ppout_reordered[ppout_reordered > pin_valid[0]] - pin_valid[0])
+        if ppout_reordered[-1] < pin_valid[-1]:  # Extrapolate below range
+            slope = (xin_valid[-1] - xin_valid[-2]) / (pin_valid[-1] - pin_valid[-2])
+            xxout[ppout_reordered < pin_valid[-1]] = xin_valid[-1] + slope * (ppout_reordered[ppout_reordered < pin_valid[-1]] - pin_valid[-1])
+
+    # Reorder output back to original if necessary
+    if ppout[0] < ppout[1]:
+        xxout = xxout[::-1]
+
+    return xxout
+
 
 @jit(nopython=True)
 def int2p(pin, xin, pout, linlog):
@@ -154,8 +178,6 @@ def int2p(pin, xin, pout, linlog):
     return xout
 
 
-
-
 def int2p_n(pin, xin, pout, linlog, dim=-1):
     """
     Wrapper for int2p to handle multi-dimensional arrays along a specific dimension.
@@ -202,9 +224,6 @@ def int2p_n(pin, xin, pout, linlog, dim=-1):
     xout = np.moveaxis(xout, -1, dim)
 
     return xout
-
-
-
 
 
 @jit(nopython=True)
@@ -262,6 +281,7 @@ def p2hyo(p_levels, data_on_p_levels, ps, a_coeff, b_coeff, p0, kflag):
 
     return data_on_hybrid_levels
 
+
 def pressure_to_hybrid(p_levels, data_on_p_levels, ps, a_coeff, b_coeff, level_dim=0, p0=100000, kflag=1):
     """
     Interpolate data on constant pressure levels to hybrid sigma levels.
@@ -274,7 +294,7 @@ def pressure_to_hybrid(p_levels, data_on_p_levels, ps, a_coeff, b_coeff, level_d
 
     # Check if p_levels is monotonically increasing
     if not np.all(np.diff(p_levels) > 0):
-        print("p_levels is not monotonically increasing. Attempting to flip arrays...")
+        logging.info("p_levels is not monotonically increasing. Attempting to flip arrays...")
 
         # Flip the p_levels and data_on_p_levels arrays
         p_levels = p_levels[::-1]
@@ -289,9 +309,10 @@ def pressure_to_hybrid(p_levels, data_on_p_levels, ps, a_coeff, b_coeff, level_d
 
     # Print elapsed time
     elapsed_time = time.time() - start_time
-    print(f"Interpolation completed in {elapsed_time:.2f} seconds.")
+    logging.info(f"Interpolation completed in {elapsed_time:.2f} seconds.")
 
     return data_on_hybrid_levels
+
 
 def pres2hyb_all(data_vars, ps, hya, hyb):
     """
@@ -300,9 +321,9 @@ def pres2hyb_all(data_vars, ps, hya, hyb):
     This function is a wrapper that prepares the input for the JIT-compiled version.
     """
 
-    allowable_interp_vars = ['t', 'u', 'v', 'q', 'cldice', 'cldliq', 'z', 'theta', 'rho', 'w']
-
     data_vint = {}
+
+    allowable_interp_vars = ['t', 'u', 'v', 'q', 'cldice', 'cldliq', 'z', 'theta', 'rho', 'w']
 
     # Create non-interpolated vars
     data_vint['hya'] = hya
@@ -317,33 +338,429 @@ def pres2hyb_all(data_vars, ps, hya, hyb):
 
     return data_vint
 
-# def pressure_to_hybrid(p_levels, data_on_p_levels, ps, a_coeff, b_coeff, p0=100000):
-#     """
-#     Interpolate data on constant pressure levels to hybrid sigma levels.
-#
-#     This function is a wrapper that prepares the input for the JIT-compiled version.
-#     """
-#     # Start timing
-#     start_time = time.time()
-#
-#     # Compute hybrid pressure levels
-#     hybrid_p = np.array([a * p0 + b * ps for a, b in zip(a_coeff, b_coeff)])
-#
-#     # Check if p_levels is monotonically increasing
-#     if not np.all(np.diff(p_levels) > 0):
-#         raise ValueError("p_levels must be monotonically increasing")
-#
-#     # Call the JIT-compiled function
-#     data_on_hybrid_levels = pressure_to_hybrid_numba(p_levels, data_on_p_levels, ps, hybrid_p)
-#
-#     # Print elapsed time
-#     elapsed_time = time.time() - start_time
-#     print(f"Interpolation completed in {elapsed_time:.2f} seconds.")
-#
-#     return data_on_hybrid_levels
-#
-#
 
+def interp_hybrid_to_pressure_wrapper(data_vars, ps, hyam, hybm, new_levels, lev_dim=0, method='log', extrapolate=True):
+
+    allowable_interp_vars = ['t', 'u', 'v', 'q', 'cldice', 'cldliq', 'z', 'theta', 'rho', 'w']
+
+    for var_name, data in data_vars.items():
+        if isinstance(data, np.ndarray) and data.ndim == 3:
+            variable_type = 'temperature' if var_name == 't' else 'geopotential' if var_name == 'z' else 'other'
+            logging.info(f"interp_hybrid_to_pressure for variable {var_name}, using {variable_type}")
+            data_vars[var_name] = interp_hybrid_to_pressure(
+                data=data,
+                ps=ps,
+                hyam=hyam,
+                hybm=hybm,
+                new_levels=new_levels,
+                lev_dim=lev_dim,
+                method=method,
+                extrapolate=extrapolate,
+                variable=variable_type,
+                t_bot=data_vars['ts'],
+                phi_sfc=data_vars['phis']
+            )
+    return data_vars
+
+
+def _temp_extrapolate(data, lev_dim, lev, p_sfc, ps, phi_sfc):
+    """Extrapolates temperature below ground using the ECMWF formulation."""
+    R_d = 287.04  # dry air gas constant
+    g_inv = 1 / 9.80616  # inverse of gravity
+    alpha = 0.0065 * R_d * g_inv
+
+    tstar = data.take(indices=-1, axis=lev_dim) * (1 + alpha * (ps / p_sfc - 1))
+    hgt = phi_sfc * g_inv
+    t0 = tstar + 0.0065 * hgt
+    tplat = np.minimum(298, t0)
+
+    tprime0 = np.where((2000 <= hgt) & (hgt <= 2500),
+                       0.002 * ((2500 - hgt) * t0 + (hgt - 2000) * tplat),
+                       np.nan)
+    tprime0 = np.where(2500 < hgt, tplat, tprime0)
+
+    alnp = np.where(hgt < 2000, alpha * np.log(lev / ps),
+                    R_d * (tprime0 - tstar) / phi_sfc * np.log(lev / ps))
+    alnp = np.where(tprime0 < tstar, 0, alnp)
+
+    result = tstar * (1 + alnp + 0.5 * (alnp**2) + (1 / 6) * (alnp**3))
+
+    return result
+
+
+def _geo_height_extrapolate(t_bot, lev, p_sfc, ps, phi_sfc):
+    """Extrapolates geopotential height below ground using the ECMWF formulation."""
+    R_d = 287.04  # dry air gas constant
+    g_inv = 1 / 9.80616  # inverse of gravity
+    alpha = 0.0065 * R_d * g_inv
+
+    tstar = t_bot * (1 + alpha * (ps / p_sfc - 1))
+    hgt = phi_sfc * g_inv
+    t0 = tstar + 0.0065 * hgt
+
+    epsilon = 1e-12
+    # Added epsilon so we don't divide by zero
+    alph = np.where((tstar <= 290.5) & (t0 > 290.5),
+                    R_d / (phi_sfc + epsilon) * (290.5 - tstar), alpha)
+
+    alph = np.where((tstar > 290.5) & (t0 > 290.5), 0, alph)
+    tstar = np.where((tstar > 290.5) & (t0 > 290.5), 0.5 * (290.5 + tstar), tstar)
+    tstar = np.where(tstar < 255, 0.5 * (tstar + 255), tstar)
+
+    alnp = alph * np.log(lev / ps)
+    return hgt - R_d * tstar * g_inv * np.log(lev / ps) * (1 + 0.5 * alnp + (1 / 6) * alnp**2)
+
+
+def _linear_interpolate_1d(new_levels, p_levels, data_on_p_levels, axis=0, fill_value=np.nan):
+    """Custom linear interpolation along a specified axis using a linear x-scale."""
+
+    # Ensure the axis of interpolation is the first axis for easier manipulation
+    p_levels = np.moveaxis(p_levels, axis, 0)
+    data_on_p_levels = np.moveaxis(data_on_p_levels, axis, 0)
+    new_levels = np.moveaxis(new_levels, axis, 0)
+
+    # Initialize output array with the shape of data_on_p_levels, but with the new_levels size
+    interp_shape = list(data_on_p_levels.shape)
+    interp_shape[0] = len(new_levels)  # Update size of the interpolation axis to match new_levels
+    interp_data = np.full(interp_shape, fill_value, dtype=data_on_p_levels.dtype)
+
+    # Perform interpolation for each slice along the axis
+    for idx in np.ndindex(data_on_p_levels.shape[1:]):
+        # Extract 1D slices for interpolation
+        p_slice = p_levels[(slice(None),) + idx]
+        dp_slice = data_on_p_levels[(slice(None),) + idx]
+
+        # Interpolate over the 1D slice
+        interp_slice = np.interp(new_levels, p_slice, dp_slice, left=fill_value, right=fill_value)
+
+        # Assign the interpolated slice back to the corresponding location in interp_data
+        interp_data[(slice(None),) + idx] = interp_slice
+
+    # Move the interpolated data back to its original axis
+    return np.moveaxis(interp_data, 0, axis)
+
+
+def _log_interpolate_1d(new_levels, p_levels, data_on_p_levels, axis=0, fill_value=np.nan):
+    """Custom log-linear interpolation along a specified axis using a logarithmic x-scale."""
+
+    # Convert pressure levels and new levels to their logarithmic scales
+    log_new_levels = np.log(new_levels)
+    log_p_levels = np.log(p_levels)
+
+    # Initialize output array with the shape of data_on_p_levels, but with the new_levels size
+    interp_shape = list(data_on_p_levels.shape)
+    interp_shape[0] = len(new_levels)  # Update size of the interpolation axis to match new_levels
+    interp_data = np.full(interp_shape, fill_value, dtype=data_on_p_levels.dtype)
+
+    # Perform interpolation for each slice along the axis
+    for idx in np.ndindex(data_on_p_levels.shape[1:]):
+        # Extract 1D slices for interpolation
+        lp_slice = log_p_levels[(slice(None),) + idx]
+        dp_slice = data_on_p_levels[(slice(None),) + idx]
+
+        #logging.info("------------")
+        #logging.info(np.exp(log_new_levels))
+        #logging.info(np.exp(lp_slice))
+        #logging.info(dp_slice)
+        # Interpolate over the 1D slice
+        interp_slice = np.interp(log_new_levels, lp_slice, dp_slice, left=-999.9, right=fill_value)
+        #logging.info(interp_slice)
+        if np.any(interp_slice == -999.9):
+            # Find the indices of the first two valid values (not equal to -999.9)
+            valid_indices = np.where(interp_slice != -999.9)[0][:2]
+
+            # Get the first two valid values
+            first_two_valid_values = interp_slice[valid_indices]
+
+            # Calculate the corresponding log pressure levels for the first two valid values
+            lp_valid = log_new_levels[valid_indices]
+
+            # Calculate the derivative (slope) between the first two valid points
+            slope = (first_two_valid_values[1] - first_two_valid_values[0]) / (lp_valid[1] - lp_valid[0])
+
+            # Extrapolate the -999.9 values using the slope
+            for i in np.where(interp_slice == -999.9)[0]:
+                interp_slice[i] = first_two_valid_values[0] + slope * (log_new_levels[i] - lp_valid[0])
+        #logging.info(interp_slice)
+        # Assign the interpolated slice back to the corresponding location in interp_data
+        interp_data[(slice(None),) + idx] = interp_slice
+
+    # Move the interpolated data back to its original axis
+    return interp_data
+
+
+def _func_interpolate(method='linear'):
+    """Define custom interpolation function."""
+    if method == 'linear':
+        return _linear_interpolate_1d
+    elif method == 'log':
+        return _log_interpolate_1d
+    else:
+        raise ValueError(f'Unknown interpolation method: {method}. Supported methods are: "log" and "linear".')
+
+
+def _pressure_from_hybrid(psfc, hya, hyb, p0=100000.):
+    """Calculate pressure at the hybrid levels."""
+    # Dynamically expand hya and hyb to match the shape of psfc
+    expanded_shape = (len(hya),) + (1,) * (psfc.ndim)
+    hya_expanded = hya.reshape(expanded_shape)
+    hyb_expanded = hyb.reshape(expanded_shape)
+    return hya_expanded * p0 + hyb_expanded * psfc
+
+
+def _vertical_remap(func_interpolate, new_levels, xcoords, data, interp_axis=0):
+    """Execute the defined interpolation function on data."""
+    with warnings.catch_warnings():
+        warnings.filterwarnings(
+            "ignore", r"Interpolation point out of data bounds encountered")
+        return func_interpolate(new_levels, xcoords, data, axis=interp_axis)
+
+
+def _vertical_remap_extrap(new_levels, lev_dim, data, output, pressure, ps, variable, t_bot, phi_sfc):
+    """A helper function to call the appropriate extrapolation function based on the user's inputs."""
+    # Get indices for the top and bottom of the vertical dimension
+    bottom_index = -1
+
+    bottom_level = np.take_along_axis(pressure, np.expand_dims(np.argmax(pressure, axis=lev_dim), axis=lev_dim), axis=lev_dim).squeeze()
+    bottom_value = data.take(indices=bottom_index, axis=lev_dim)
+
+    # Prepare new_levels for broadcasting
+    new_levels_broadcasted = np.expand_dims(new_levels, axis=tuple(range(1, pressure.ndim)))
+
+    # Downward Extrapolation
+    if variable == 'temperature':
+        extrapolated = _temp_extrapolate(data, lev_dim, new_levels_broadcasted, bottom_level, ps, phi_sfc)
+    elif variable == 'geopotential':
+        extrapolated = _geo_height_extrapolate(t_bot, new_levels_broadcasted, bottom_level, ps, phi_sfc)
+    else:
+        extrapolated = bottom_value
+
+    # Apply downward extrapolation
+    output = np.where(new_levels_broadcasted > (bottom_level - 0.1), extrapolated, output)
+
+    # Debugging: Count and print number of NaNs after downward extrapolation
+    nan_count_downward = np.isnan(output).sum()
+    if nan_count_downward > 0:
+        logging.info(f"NaNs detected after downward extrapolation: {nan_count_downward}")
+        nan_indices_downward = np.argwhere(np.isnan(output))
+        logging.info("NaNs found at indices (downward extrapolation):")
+        logging.info(nan_indices_downward)
+        logging.info("Inspecting inputs at NaN indices for downward extrapolation:")
+        for idx in nan_indices_downward:
+            logging.info(f"Index {idx}:")
+
+            # Ensure the index is within bounds for the data array
+            if idx[0] < data.shape[0] and all(i < s for i, s in zip(idx[1:], data.shape[1:])):
+                logging.info(f"data at index: {data[tuple(idx)]}")
+            else:
+                logging.info("Data at index: Index out of bounds for data array.")
+
+            # Ensure the index is within bounds for the bottom_level, ps, and phi_sfc arrays
+            if all(i < s for i, s in zip(idx[1:], bottom_level.shape)):
+                logging.info(f"bottom_level at index: {bottom_level[tuple(idx[1:])]}")
+                logging.info(f"ps at index: {ps[tuple(idx[1:])]}")
+                logging.info(f"phi_sfc at index: {phi_sfc[tuple(idx[1:])]}")
+            else:
+                logging.info("Bottom_level, ps, or phi_sfc index out of bounds.")
+
+            # Print the corresponding new_levels value
+            if idx[0] < len(new_levels):
+                logging.info(f"new_levels at level index: {new_levels[idx[0]]}")
+            else:
+                logging.info("new_levels index out of bounds.")
+
+    return output
+
+
+def interp_hybrid_to_pressure(data: np.ndarray,
+                              ps: np.ndarray,
+                              hyam: np.ndarray,
+                              hybm: np.ndarray,
+                              p0: float = 100000.,
+                              new_levels: np.ndarray = __pres_lev_mandatory__,
+                              lev_dim: int = 0,
+                              method: str = 'linear',
+                              extrapolate: bool = False,
+                              variable: str = None,
+                              t_bot: np.ndarray = None,
+                              phi_sfc: np.ndarray = None) -> np.ndarray:
+    """Interpolate and extrapolate data from hybrid-sigma levels to isobaric levels."""
+
+    # Check inputs
+    if extrapolate and variable is None:
+        raise ValueError("If extrapolate is True, variable must be provided.")
+    if variable in ['geopotential', 'temperature'] and (t_bot is None or phi_sfc is None):
+        raise ValueError("If variable is 'geopotential' or 'temperature', both t_bot and phi_sfc must be provided")
+    if variable not in ['geopotential', 'temperature', 'other', None]:
+        raise ValueError(f"The value of variable is {variable}, but the accepted values are 'temperature', 'geopotential', 'other', or None.")
+
+    func_interpolate = _func_interpolate(method)
+
+    # Calculate pressure levels at the hybrid levels
+    pressure = _pressure_from_hybrid(ps, hyam, hybm, p0)  # Pa
+
+    # Prepare output data structure
+    out_shape = list(data.shape)
+    out_shape[lev_dim] = new_levels.size
+    output = np.empty(out_shape, dtype=data.dtype)
+
+    # Perform the vertical interpolation
+    output = _vertical_remap(
+        func_interpolate, new_levels, pressure, data, interp_axis=lev_dim)
+
+    if extrapolate:
+        output = _vertical_remap_extrap(new_levels, lev_dim, data, output, pressure, ps, variable, t_bot, phi_sfc)
+
+    return output
+
+
+def interpolate_mpas_columns_wrapper(mpas_data, data_horiz):
+
+    t_wrf = np.zeros((mpas_data['nlev'], mpas_data['ncell']))
+    theta_wrf = np.zeros((mpas_data['nlev'], mpas_data['ncell']))
+    rho_wrf = np.zeros((mpas_data['nlev'], mpas_data['ncell']))
+    w_wrf = np.zeros((mpas_data['nlevi'], mpas_data['ncell']))
+    q_wrf = np.zeros((mpas_data['nlev'], mpas_data['ncell']))
+    u_wrf = np.zeros((mpas_data['nlev'], mpas_data['ncell']))
+    v_wrf = np.zeros((mpas_data['nlev'], mpas_data['ncell']))
+
+    for ix in range(mpas_data['ncell']):
+        if ix % 10000 == 0:
+            logging.info(f"... MPAS_VERT_INTERP: {100. * ix / (mpas_data['ncell'] - 1):.2f}%")
+
+        if ix == 0:
+            z_src_orientation = 'bottom_to_top' if data_horiz['z'][-1, ix] > data_horiz['z'][0, ix] else 'top_to_bottom'
+            z_mpas_orientation = 'bottom_to_top' if mpas_data['z'][ix, -1] > mpas_data['z'][ix, 0] else 'top_to_bottom'
+
+            logger.debug(f"z_src_orientation is {z_src_orientation}")
+            logger.debug(f"z_mpas_orientation is {z_mpas_orientation}")
+
+            # Check if the orientations are the same or different
+            if z_src_orientation == z_mpas_orientation:
+                logger.debug("z_fv and mpas_z have the same orientation")
+                flip_array = False
+            else:
+                logger.debug("z_fv and mpas_z have different orientations")
+                logger.debug("we will flip arrays in interpolate_mpas_columns_wrapper")
+                flip_array = True
+
+        # We want to priorize the mpas grid orientation, so flip the src arrays if z
+        # orientation differs between the two
+        if flip_array:
+            t_wrf[:, ix], theta_wrf[:, ix], rho_wrf[:, ix], w_wrf[:, ix], \
+            q_wrf[:, ix], u_wrf[:, ix], v_wrf[:, ix] = interpolate_single_mpas_column(
+                ix,
+                mpas_data['nlev'], mpas_data['nlevi'], mpas_data['z'],
+                data_horiz['t'][::-1, ix], data_horiz['z'][::-1, ix],
+                data_horiz['theta'][::-1, ix], data_horiz['rho'][::-1, ix],
+                data_horiz['w'][::-1, ix], data_horiz['q'][::-1, ix],
+                data_horiz['u'][::-1, ix], data_horiz['v'][::-1, ix]
+            )
+        else:
+            t_wrf[:, ix], theta_wrf[:, ix], rho_wrf[:, ix], w_wrf[:, ix], \
+            q_wrf[:, ix], u_wrf[:, ix], v_wrf[:, ix] = interpolate_single_mpas_column(
+                ix,
+                mpas_data['nlev'], mpas_data['nlevi'], mpas_data['z'],
+                data_horiz['t'][:, ix], data_horiz['z'][:, ix],
+                data_horiz['theta'][:, ix], data_horiz['rho'][:, ix],
+                data_horiz['w'][:, ix], data_horiz['q'][:, ix],
+                data_horiz['u'][:, ix], data_horiz['v'][:, ix]
+            )
+
+    data_horiz['t'] = t_wrf
+    data_horiz['theta'] = theta_wrf
+    data_horiz['rho'] = rho_wrf
+    data_horiz['w'] = w_wrf
+    data_horiz['q'] = q_wrf
+    data_horiz['u'] = u_wrf
+    data_horiz['v'] = v_wrf
+
+    return data_horiz
+
+
+def interpolate_single_mpas_column(ix, mpas_nlev, mpas_nlevi, mpas_z, t_fv, z_fv, theta_fv, rho_fv, w_fv, q_fv, u_fv, v_fv):
+
+    zmid = (mpas_z[ix, 1:mpas_nlevi] + mpas_z[ix, 0:mpas_nlevi - 1]) / 2.0
+    zint = mpas_z[ix, :]
+
+    t_wrf_col = np.zeros(mpas_nlev)
+    theta_wrf_col = np.zeros(mpas_nlev)
+    rho_wrf_col = np.zeros(mpas_nlev)
+    w_wrf_col = np.zeros(mpas_nlevi)
+    q_wrf_col = np.zeros(mpas_nlev)
+    u_wrf_col = np.zeros(mpas_nlev)
+    v_wrf_col = np.zeros(mpas_nlev)
+
+    if ix == 0:
+        logger.debug("interpolate_single_mpas_column T metrics at ix = 0:")
+        logger.debug(f"z_fv[0]: {z_fv[0]}, z_fv[nlev-1]: {z_fv[-1]}")
+        logger.debug(f"zmid[0]: {zmid[0]}, zmid[nlev-1]: {zmid[-1]}")
+        logger.debug(f"t_fv[0]: {t_fv[0]}, t_fv[nlev-1]: {t_fv[-1]}")
+        logger.debug(f"t_fv range: {t_fv.min()} to {t_fv.max()}")
+        logger.debug(f"t_fv size: {t_fv.size}, z_fv size: {z_fv.size}, zmid size: {zmid.size}")
+
+    t_wrf_col = z_to_z_interp(t_fv, z_fv, zmid, extrapLow=True, extrapHigh=True)
+
+    theta_wrf_col = z_to_z_interp(theta_fv, z_fv, zmid, extrapLow=True, extrapHigh=True)
+    rho_wrf_col = z_to_z_interp(rho_fv, z_fv, zmid, extrapLow=True, extrapHigh=True)
+    w_wrf_col = z_to_z_interp(w_fv, z_fv, zint, extrapLow=True, extrapHigh=True)
+    q_wrf_col = z_to_z_interp(q_fv, z_fv, zmid, extrapLow=True, extrapHigh=True)
+
+    # u and v we don't extrapolate aloft to prevent wind speeds from getting too high
+    u_wrf_col = z_to_z_interp(u_fv, z_fv, zmid, extrapLow=True, extrapHigh=False)
+    v_wrf_col = z_to_z_interp(v_fv, z_fv, zmid, extrapLow=True, extrapHigh=False)
+
+#     if ix == 30:
+#       logging.info(u_wrf_col)
+#       logging.info(u_fv[::-1])
+#       logging.info(zmid)
+#       logging.info(z_fv[::-1])
+
+    return t_wrf_col, theta_wrf_col, rho_wrf_col, w_wrf_col, q_wrf_col, u_wrf_col, v_wrf_col
+
+
+@jit(nopython=True)
+def z_to_z_interp(theta_fv, z_fv, thisCol, extrapLow=False, extrapHigh=False):
+    nlev = len(thisCol)
+    t_wrf = np.zeros(nlev)
+
+    for i in range(nlev):
+        # logger.debug(f"thisCol[i]: {thisCol[i]}, z_fv[0] {z_fv[0]}, z_fv[-1], {z_fv[-1]}")
+        if thisCol[i] <= z_fv[0] or thisCol[i] >= z_fv[-1]:
+            t_wrf[i] = np.nan
+        else:
+            for j in range(len(z_fv) - 1):
+                if z_fv[j] <= thisCol[i] <= z_fv[j + 1]:
+                    t_wrf[i] = theta_fv[j] + (theta_fv[j + 1] - theta_fv[j]) * (thisCol[i] - z_fv[j]) / (z_fv[j + 1] - z_fv[j])
+                    break
+
+#     if np.all(np.isnan(t_wrf)):
+#         logging.error("Interpolation failed: t_wrf array is entirely NaN. Exiting program.")
+#         logging.error("This *may* be because the arrays you have sent in are flipped...")
+#         sys.exit(1)
+
+    if np.any(np.isnan(t_wrf)):
+        ixvalid = np.where(~np.isnan(t_wrf))[0]
+
+        if np.isnan(t_wrf[0]):
+            lowest_valid = np.min(ixvalid)
+            if extrapLow:
+                deriv = (t_wrf[lowest_valid] - t_wrf[lowest_valid + 1]) / (thisCol[lowest_valid] - thisCol[lowest_valid + 1])
+            else:
+                deriv = 0.0
+            for ff in range(0, lowest_valid):
+                t_wrf[ff] = t_wrf[lowest_valid] + deriv * (thisCol[ff] - thisCol[lowest_valid])
+
+        if np.isnan(t_wrf[-1]):
+            highest_valid = np.max(ixvalid)
+            if extrapHigh:
+                deriv = (t_wrf[highest_valid - 1] - t_wrf[highest_valid]) / (thisCol[highest_valid - 1] - thisCol[highest_valid])
+            else:
+                deriv = 0.0
+            for ff in range(highest_valid + 1, nlev):
+                t_wrf[ff] = t_wrf[highest_valid] + deriv * (thisCol[ff] - thisCol[highest_valid])
+
+    return t_wrf
 
 
 ###### NEEDS TO BE TESTED
@@ -372,11 +789,11 @@ def ds2hbd(dati, sigi, sigo, intyp):
     for k in range(nlo):
         if sigo[k] <= sigi[1]:
             kp = 0
-        elif sigo[k] >= sigi[nli-2]:
+        elif sigo[k] >= sigi[nli - 2]:
             kp = nli - 2
         else:
             kp = 0
-            while sigo[k] > sigi[kp+1]:
+            while sigo[k] > sigi[kp + 1]:
                 kp += 1
 
         if intyp == 1:  # Linear interpolation
@@ -387,6 +804,7 @@ def ds2hbd(dati, sigi, sigo, intyp):
             dato[k] = dati[kp] + (dati[kp+1] - dati[kp]) * (a2ln(sigo[k]) - a2ln(sigi[kp])) / (a2ln(sigi[kp+1]) - a2ln(sigi[kp]))
 
     return dato
+
 
 @jit(nopython=True)
 def dh2sdrv(dati, hya, hyb, p0, psfc, sigi, intyp):
@@ -405,7 +823,6 @@ def dh2sdrv(dati, hya, hyb, p0, psfc, sigi, intyp):
     Returns:
     - dato: Interpolated data on hybrid coordinates (1D array)
     """
-    nlvi = sigi.size
     nlvo = hya.size
     sigo = np.zeros(nlvo)
 
@@ -430,13 +847,13 @@ def sigma_to_hybrid(dati, sigma, hya, hyb, psfc, p0=100000, intyp=1):
 
     # Check if sigma is monotonically increasing
     if not np.all(np.diff(sigma) > 0):
-       raise ValueError("Sigma levels must be monotonically increasing")
+        raise ValueError("Sigma levels must be monotonically increasing")
 
     # Call the Python-equivalent Fortran function
     dato = dh2sdrv(dati, hya, hyb, p0, psfc, sigma, intyp)
 
     # Print elapsed time
     elapsed_time = time.time() - start_time
-    print(f"Sigma to hybrid interpolation completed in {elapsed_time:.2f} seconds.")
+    logging.info(f"Sigma to hybrid interpolation completed in {elapsed_time:.2f} seconds.")
 
     return dato
